♻️refactor: refactor solicitation controllers
