+++ src/main/java/br/com/chronos/server/api/controllers/solicitation/justification_type/JustificationTypeController.java
@@ -1,4 +1,4 @@
-package br.com.chronos.server.api.controllers.solicitation;
+package br.com.chronos.server.api.controllers.solicitation.justification_type;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -11,7 +11,5 @@ import org.springframework.web.bind.annotation.RestController;
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @RestController
-@RequestMapping("/solicitation")
-public @interface SolicitationsController {
-
-}
+@RequestMapping("solicitation/justification-type")
+public @interface JustificationTypeController {}
