+++ src/main/java/br/com/chronos/server/api/controllers/solicitation/justification_type/ListAllJustificationTypeController.java
@@ -1,4 +1,4 @@
-package br.com.chronos.server.api.controllers.solicitation;
+package br.com.chronos.server.api.controllers.solicitation.justification_type;
 
 import java.util.List;
 
@@ -9,16 +9,16 @@ import org.springframework.web.bind.annotation.GetMapping;
 
 import br.com.chronos.core.solicitation.domain.dtos.JustificationTypeDto;
 import br.com.chronos.core.solicitation.interfaces.repositories.JustificationTypeRepository;
-import br.com.chronos.core.solicitation.use_cases.ListAllJustificationTypeUseCase;
+import br.com.chronos.core.solicitation.use_cases.ListAllJustificationTypesUseCase;
 
-@SolicitationsController
+@JustificationTypeController
 public class ListAllJustificationTypeController {
   @Autowired
   private JustificationTypeRepository justificationTypeRepository;
 
   @GetMapping("/justification-type")
   public ResponseEntity<List<JustificationTypeDto>> handle() {
-    var useCase = new ListAllJustificationTypeUseCase(justificationTypeRepository);
+    var useCase = new ListAllJustificationTypesUseCase(justificationTypeRepository);
     var response = useCase.execute();
     return ResponseEntity.status(HttpStatus.OK).body(response);
   }
