+++ src/main/java/br/com/chronos/server/api/controllers/solicitation/justification_type/UpdateJustificationTypeController.java
@@ -1,4 +1,4 @@
-package br.com.chronos.server.api.controllers.solicitation;
+package br.com.chronos.server.api.controllers.solicitation.justification_type;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -10,7 +10,7 @@ import br.com.chronos.core.solicitation.domain.dtos.JustificationTypeDto;
 import br.com.chronos.core.solicitation.interfaces.repositories.JustificationTypeRepository;
 import br.com.chronos.core.solicitation.use_cases.UpdateJustificationTypeUseCase;
 
-@SolicitationsController
+@JustificationTypeController
 public class UpdateJustificationTypeController {
 
   @Autowired
