+++ src/main/java/br/com/chronos/server/api/controllers/solicitation/solicitations/SolicitationsController.java
@@ -1,0 +1,15 @@
+package br.com.chronos.server.api.controllers.solicitation.solicitations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@RestController
+@RequestMapping("/solicitation/solicitations")
+public @interface SolicitationsController {}
